backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
requests 
python-dotenv 
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi 
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests